Extract ModalCoverModifier and reuse in NavigationStackFlow
Pulls the sheet/full-screen cover binding boilerplate out of TabViewFlow
into a standalone ModalCoverModifier generic over any Coordinator. Apply
it in NavigationStackFlow too so all flow containers share the same
modal presentation logic.
