+++ Sources/FuturedArchitecture/Architecture/TabViewFlow.swift
@@ -63,39 +63,3 @@ public struct TabContentFlow<Coordinator: TabCoordinator, Content: TabContent<Co
         .modifier(ModalCoverModifier(coordinator: coordinator))
     }
 }
-
-// MARK: - Shared Modal Presentation
-
-/// Shared modifier that adds sheet and full-screen cover presentation to any coordinator-backed view.
-struct ModalCoverModifier<C: Coordinator>: ViewModifier {
-    var coordinator: C
-
-    private var sheetBinding: Binding<C.Destination?> {
-        .init {
-            coordinator.modalCover?.style == .sheet ? coordinator.modalCover?.destination : nil
-        } set: { destination in
-            coordinator.modalCover = destination.map { .init(destination: $0, style: .sheet) }
-        }
-    }
-
-    #if os(macOS)
-    func body(content: Content) -> some View {
-        content
-            .sheet(item: sheetBinding, onDismiss: coordinator.onModalDismiss, content: coordinator.scene(for:))
-    }
-    #else
-    private var fullscreenCoverBinding: Binding<C.Destination?> {
-        .init {
-            coordinator.modalCover?.style == .fullscreenCover ? coordinator.modalCover?.destination : nil
-        } set: { destination in
-            coordinator.modalCover = destination.map { .init(destination: $0, style: .fullscreenCover) }
-        }
-    }
-
-    func body(content: Content) -> some View {
-        content
-            .sheet(item: sheetBinding, onDismiss: coordinator.onModalDismiss, content: coordinator.scene(for:))
-            .fullScreenCover(item: fullscreenCoverBinding, onDismiss: coordinator.onModalDismiss, content: coordinator.scene(for:))
-    }
-    #endif
-}
